Update WorkManager schedule logic
