+++ app/src/main/java/com/honorida/activities/main/MainActivity.kt
@@ -1,15 +1,69 @@
 package com.honorida.activities.main
 
+import android.Manifest
+import android.app.NotificationChannel
+import android.app.NotificationManager
+import android.content.Context
+import android.content.Intent
+import android.os.Build
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.annotation.RequiresApi
+import androidx.work.Data
+import androidx.work.OneTimeWorkRequestBuilder
+import com.honorida.HonoridaApp
+import com.honorida.R
 import com.honorida.activities.main.ui.components.App
+import com.honorida.domain.constants.APP_UPDATES_NOTIFICATION_CHANNEL_ID
+import com.honorida.workers.AppUpdateWorker
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            scheduleAppUpdatesCheck()
+            createNotificationChannels()
+        }
+
         setContent {
             App()
         }
+    }
+
+    @RequiresApi(Build.VERSION_CODES.O)
+    private fun scheduleAppUpdatesCheck() {
+        val intent = intent
+        if (intent.action == Intent.ACTION_MAIN){
+            val appUpdateWorker = OneTimeWorkRequestBuilder<AppUpdateWorker>()
+                .build()
+            HonoridaApp.appModule.workManager.enqueue(appUpdateWorker)
+        }
+    }
+
+    @RequiresApi(Build.VERSION_CODES.O)
+    private fun createNotificationChannels() {
+        createNotificationChannel(
+            APP_UPDATES_NOTIFICATION_CHANNEL_ID,
+            getString(R.string.app_update_notifications),
+            description =
+            getString(R.string.notifications_of_a_new_app_version_available),
+            NotificationManager.IMPORTANCE_DEFAULT
+        )
+    }
+
+    @RequiresApi(Build.VERSION_CODES.O)
+    private fun createNotificationChannel(
+        id: String,
+        name: String,
+        description: String,
+        importance: Int
+    ) {
+        val channel = NotificationChannel(id, name, importance)
+        channel.description = description
+        val notificationManager: NotificationManager =
+            getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+        notificationManager.createNotificationChannel(channel)
     }
 }
